backend/requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart

supabase
pypdf

langchain
langchain-core
langchain-text-splitters
langchain-google-genai
google-generativeai
